+++ skops/io/exceptions.py
@@ -1,10 +1,5 @@
 class UnsupportedTypeException(TypeError):
-    """Raise when an object of this type is known to be unsupported"""
-
-    def __init__(self, obj):
-        super().__init__(
-            f"Objects of type {obj.__class__.__name__} are not supported yet."
-        )
+    """Raised when an object cannot be persisted; the message says why."""
 
 
 # Types that are known to be unsafe to trust by default, together with a